Split pipelines into categories
