+++ paz/pipelines/__init__.py
@@ -1,0 +1,20 @@
+from .image import AugmentImage
+from .image import PreprocessImage
+from .image import AutoEncoderPredictor
+from .image import EncoderPredictor
+from .image import DecoderPredictor
+
+from .detection import AugmentBoxes
+from .detection import PreprocessBoxes
+from .detection import AugmentDetection
+from .detection import SingleShotInference
+
+from .keypoints import KeypointSharedAugmentation
+from .keypoints import KeypointNetInference
+from .keypoints import PredictKeypoints
+from .keypoints import KeypointInference
+
+from .renderer import RenderTwoViews
+from .renderer import RandomizeRenderedImage
+
+from .pose import PredictPoseFromKeypoints
